Add files via upload
The chat program I'll create will be more like a chat room than a peer-to-peer chat. As a result, numerous users can connect to the chat server and send messages. any message is sent to any chat user who is connected. The theory is as basic as the construction.

The code is made up of two Python programs. The server comes first, followed by the chat client.